rnd refactoring
