+++ hoot-core/src/main/cpp/hoot/core/cmd/InfoCmd.cpp
@@ -30,8 +30,6 @@
 #include <hoot/core/info/ApiEntityDisplayInfo.h>
 #include <hoot/core/info/ConfigOptionsDisplayer.h>
 #include <hoot/core/info/FormatsDisplayer.h>
-#include <hoot/core/language/LanguageInfoProvider.h>
-#include <hoot/core/language/HootServicesLanguageInfoResponseParser.h>
 #include <hoot/core/util/Factory.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/core/util/ConfigOptions.h>
@@ -212,76 +210,6 @@ public:
           displayOutputs, displayOutputsSupportingStreaming, displayOgrOnly)
         .toStdString();
     }
-    else if (specifiedOpts.contains("--languages"))
-    {
-      args.removeAt(args.indexOf("--languages"));
-      if (args.size() != 1)
-      {
-        std::cout << getHelp() << std::endl << std::endl;
-        throw IllegalArgumentException(
-          QString("%1 with the --languages option takes one parameter.").arg(getName()));
-      }
-
-      // only allowing one option per command
-      const QStringList supportedOpts = _getSupportedLanguageOptions();
-      QStringList specifiedOpts;
-      for (int i = 0; i < args.size(); i++)
-      {
-        const QString arg = args.at(i);
-        if (specifiedOpts.contains(arg) ||
-            (supportedOpts.contains(arg) && specifiedOpts.size() > 0))
-        {
-          std::cout << getHelp() << std::endl << std::endl;
-          throw IllegalArgumentException(QString("%1 takes a single option.").arg(getName()));
-        }
-        specifiedOpts.append(arg);
-      }
-      if (specifiedOpts.size() == 0)
-      {
-        std::cout << getHelp() << std::endl << std::endl;
-        throw IllegalArgumentException(
-          QString("%1 with the --languages option takes a single option.").arg(getName()));
-      }
-      LOG_VARD(specifiedOpts.size());
-
-      ConfigOptions opts = ConfigOptions(conf());
-
-      try
-      {
-        std::shared_ptr<LanguageInfoProvider> client;
-        client.reset(
-          Factory::getInstance().constructObject<LanguageInfoProvider>(
-            opts.getLanguageInfoProvider()));
-        client->setConfiguration(conf());
-
-        const QString type = args[0].replace("--", "").toLower();
-
-        QString displayStr;
-        if (type == "translatable" || type == "detectable")
-        {
-          displayStr =
-            HootServicesLanguageInfoResponseParser::parseAvailableLanguagesResponse(
-              type, client->getAvailableLanguages(type));
-        }
-        else
-        {
-          displayStr =
-            HootServicesLanguageInfoResponseParser::parseAvailableAppsResponse(
-              type, client->getAvailableApps(type));
-        }
-        std::cout << displayStr << std::endl;
-      }
-      catch (const HootException& e)
-      {
-        LOG_VART(e.getWhat());
-        if (e.getWhat().contains("Access tokens for user"))
-        {
-          std::cout <<
-            "You must log in to the Hootenanny Web Services before displaying supported language information." <<
-            std::endl;
-        }
-      }
-    }
     else if (specifiedOpts.contains("--cleaning-operations"))
     {
       std::cout << ApiEntityDisplayInfo::getDisplayInfoOps("map.cleaner.transforms").toStdString();
@@ -328,16 +256,6 @@ public:
 
 private:
 
-  QStringList _getSupportedLanguageOptions() const
-  {
-    QStringList options;
-    options.append("--detectable");
-    options.append("--detectors");
-    options.append("--translatable");
-    options.append("--translators");
-    return options;
-  }
-
   QStringList _getSupportedOptions() const
   {
     QStringList options;
@@ -351,7 +269,6 @@ private:
     options.append("--filters");
     options.append("--formats");
     options.append("--geometry-type-criteria");
-    options.append("--languages");
     options.append("--matchers");
     options.append("--match-creators");
     options.append("--mergers");
