+++ test-files/cmd/slow/InfoCmdTest.sh
@@ -83,24 +83,6 @@ echo "Listing input formats supported bounded reads that are read by OGR..."
 echo ""
 hoot info $CONFIG --formats --input-bounded --ogr # none currently exist
 
-# LANGUAGES
-echo "Listing detectable languages..."
-echo ""
-hoot info $CONFIG -D language.info.provider=hoot::HootServicesLanguageInfoMockClient --languages --detectable
-echo ""
-echo "Listing translatable languages..."
-echo ""
-hoot info $CONFIG -D language.info.provider=hoot::HootServicesLanguageInfoMockClient --languages --translatable
-echo ""
-echo "Listing available language detectors..."
-echo ""
-hoot info $CONFIG -D language.info.provider=hoot::HootServicesLanguageInfoMockClient --languages --detectors
-echo ""
-echo "Listing available to English language translators..."
-echo ""
-hoot info $CONFIG -D language.info.provider=hoot::HootServicesLanguageInfoMockClient --languages --translators
-echo ""
-
 # MATCH CREATORS
 echo "Listing a match creator..."
 echo ""
